improved expand collapse of toolbar
now there are plus / minus icons instead of a collapse button... canvas
gets bigger when toolbar is minimized...smaller when toolbar is
maximized.

currently plus button is over top of corner of canvas.... drawing gets a
little strange in that corner... its something we can work on..when plus
is activated .... toolbar pops out and moves canvas over to the
right...its not animated...so the transition is instant but we can try
to figure out how to make animated like the drop down menu is when you
get to a narrow page width.